pyproject.toml
[build-system]
requires = ["setuptools >= 70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openanalytics"
description = "Open Analytics"
version = "1.0.0"
authors = [
    { name = "Mirai Kumiko", email = "miraikumiko@disroot.org" }
]
maintainers = [
    { name = "Mirai Kumiko", email = "miraikumiko@disroot.org" }
]
keywords = [
    "Analytics"
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.21.0",
    "alembic==1.15.1",
    "anyio==4.8.0",
    "APScheduler==3.11.0",
    "asyncpg==0.30.0",
    "build==1.2.2.post1",
    "click==8.1.8",
    "databases==0.9.0",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpx==0.28.1",
    "idna==3.10",
    "Mako==1.3.9",
    "MarkupSafe==3.0.2",
    "packaging==24.2",
    "pyproject_hooks==1.2.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.39",
    "SQLAlchemy-Utils==0.41.2",
    "starlette==0.46.1",
    "typing_extensions==4.12.2",
    "user-agents==2.2.0",
    "uvicorn==0.34.0"
]
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
Source = "https://github.com/miraikumiko/openanalytics"

[project.optional-dependencies]
dev = [
    "pytest-asyncio==0.25.0",
    "pytest-cov==6.0.0",
    "pylint==3.3.0"
]

[project.scripts]
openanalytics = "openanalytics.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "openanalytics",
    "openanalytics.*"
]
exclude = [
    "contrib"
]

[tool.pytest.ini_options]
pythonpath = [".", "openanalytics"]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope="session"
addopts = [
    "--cov=openanalytics"
]

[tool.pylint]
disable = ["C0114", "C0115", "C0116"]
max-line-length = 200